requirements.txt
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-core
langchain
python-dotenv
streamlit
                                                